Изменил в Animal поле age с byte на int

+++ src/home_work_5/comparators/AgeOfAnimalsAndNickComparator.java
@@ -7,8 +7,8 @@ import java.util.Comparator;
 public class AgeOfAnimalsAndNickComparator implements Comparator<Animal> {
     @Override
     public int compare(Animal a1, Animal a2) {
-        byte age1 = a1.getAge();
-        byte age2 = a2.getAge();
+        int age1 = a1.getAge();
+        int age2 = a2.getAge();
 
         if (age1 != age2) {
             return age1 - age2;
